+++ src/classic/a2dp_sink.h
@@ -128,6 +128,15 @@ uint8_t a2dp_sink_establish_stream(bd_addr_t remote, uint8_t local_seid, uint16_
 void a2dp_sink_disconnect(uint16_t a2dp_cid);
 
 /**
+ * @brief Register media configuration validator. Can reject insuitable configuration or report stream endpoint as currently busy
+ * @note validator has to return AVDTP error codes like: AVDTP_ERROR_CODE_SEP_IN_USE or AVDTP_ERROR_CODE_UNSUPPORTED_CONFIGURATION
+ *       the callback receives the media configuration in the same format as the existing A2dP_SUBEVENT_SIGNALING_MEDIA_CODEC_SBC_CONFIGURATION
+ *       and similar
+ * @param callback
+ */
+void a2dp_sink_register_media_config_validator(uint8_t (*callback)(const avdtp_stream_endpoint_t * stream_endpoint, const uint8_t * event, uint16_t size));
+
+/**
  * @brief De-Init A2DP Sink device.
  */
 void a2dp_sink_deinit(void);